:building_construction: add eventBus.js - move file summrizaiton to eventBus - add task when file set to public
